requirements.txt
Django>=3.1.2,<3.2.0
djangorestframework>=3.12.1,<3.13
psycopg2>=2.8.6,<2.9
Pillow>=7.2.0,<7.3
django-cors-headers==3.11.0
flake8>=3.8.4,<3.9
